Add database

+++ src/main/kotlin/models/AbstractProfileBackend.kt
@@ -3,9 +3,9 @@ package models
 import models.profile.*
 
 interface AbstractProfileBackend {
-    fun postStudentProfile(id: Int, profile: StudentProfile)
-    fun getStudentProfile(id: Int): StudentProfile
+    fun postStudentProfile(id: Long, profile: StudentProfile)
+    fun getStudentProfile(id: Long): StudentProfile
 
-    fun postInstructorProfile(id: Int, profile: InstructorProfile)
-    fun getInstructorProfile(id: Int): InstructorProfile
+    fun postInstructorProfile(id: Long, profile: InstructorProfile)
+    fun getInstructorProfile(id: Long): InstructorProfile
 }
